requirements.txt
flask
flask-restx
flask-cors
gunicorn
werkzeug
pymongo[srv]
python-dotenv
python-jose
six